split create instance dialog pages
